Refactor: 장바구니 페이지 유틸 함수 로직 분리 및 절대 경로 수정
