+++ index.html
@@ -5,10 +5,10 @@
       type="text/javascript"
       src="//dapi.kakao.com/v2/maps/sdk.js?appkey=2b7174d397f6532d16f44c8233be8d4b&libraries=services"
     ></script>
-    <meta
+    <!-- <meta
       http-equiv="Content-Security-Policy"
       content="upgrade-insecure-requests"
-    />
+    /> -->
     <meta charset="UTF-8" />
     <link rel="icon" type="image/svg+xml" href="/fastcatch.svg" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
