+++ src/components/commonFilter/filterDropdowns/AmountDropdown.tsx
@@ -3,7 +3,7 @@ import { FaPlus } from "react-icons/fa";
 import { FaMinus } from "react-icons/fa";
 
 import { useRecoilState } from "recoil";
-import { filterState } from "@/src/states/filterState";
+import { filterState } from "@/states/filterState";
 
 interface dropdownProps {
   isSelected: "location" | "date" | "amount" | null;
@@ -16,7 +16,7 @@ const AmountDropdown = (props: dropdownProps) => {
     event.stopPropagation();
     if (filterStates.amount === 15 && arg === 1) return; //15명인 상황에서 +1을 하면 바로 return
     if (filterStates.amount === 1 && arg === -1) return; //1명인 상황에서 -1을 하면 바로 return
-    setFilterStates((prevStates) => ({
+    setFilterStates(prevStates => ({
       ...prevStates,
       amount: prevStates.amount + arg,
     }));
@@ -30,14 +30,14 @@ const AmountDropdown = (props: dropdownProps) => {
           <div className="amountChangeContainer">
             <div //
               className="amountChangeButton"
-              onClick={(event) => amountChangeHander(event, 1)}
+              onClick={event => amountChangeHander(event, 1)}
             >
               <FaPlus />
             </div>
             <div className="text-body2">{filterStates.amount}명</div>
             <div //
               className="amountChangeButton"
-              onClick={(event) => amountChangeHander(event, -1)}
+              onClick={event => amountChangeHander(event, -1)}
             >
               <FaMinus />
             </div>
