+++ src/components/commonFilter/filterDropdowns/locationDropdownEl/LocationCategoryName.tsx
@@ -2,12 +2,14 @@ import React from "react";
 import "./locationDropdownEls.scss";
 
 import { useRecoilState } from "recoil";
-import { filterState } from "@/src/states/filterState";
-import { filterStateTypes } from "@/src/states/filterState";
+import { filterState } from "@/states/filterState";
+import { filterStateTypes } from "@/states/filterState";
 
 interface locationProps {
   allocated: [string, string];
-  onChangeSelected: React.Dispatch<React.SetStateAction<"location" | "date" | "amount" | null>>;
+  onChangeSelected: React.Dispatch<
+    React.SetStateAction<"location" | "date" | "amount" | null>
+  >;
 }
 
 const LocationCategoryName = (props: locationProps) => {
@@ -15,8 +17,11 @@ const LocationCategoryName = (props: locationProps) => {
 
   const selectionHandler = (event: React.MouseEvent) => {
     event.stopPropagation(); // 버블링 방지
-    setFilterStates((prevStates) => {
-      return { ...prevStates, locale: props.allocated[0] as filterStateTypes["locale"] };
+    setFilterStates(prevStates => {
+      return {
+        ...prevStates,
+        locale: props.allocated[0] as filterStateTypes["locale"],
+      };
     });
 
     props.onChangeSelected("date"); // 일정 필터로 바꿈
@@ -24,7 +29,9 @@ const LocationCategoryName = (props: locationProps) => {
 
   return (
     <div
-      className={props?.allocated[0] === filterStates.locale ? "selected" : "unSelected"} //
+      className={
+        props?.allocated[0] === filterStates.locale ? "selected" : "unSelected"
+      } //
       onClick={selectionHandler}
     >
       {props.allocated[1]}
