+++ src/pages/accommodation/roomSelect/roomInfo/RoomInfo.tsx
@@ -1,14 +1,14 @@
-import instance from "@/src/api/instanceApi";
-import CommonBadge from "@/src/components/commonBadge/CommonBadge";
-import CommonButton from "@/src/components/commonButton/CommonButton";
-import CommonToastLayout from "@/src/components/commonToast/CommonToastLayout";
-import { filterState } from "@/src/states/filterState";
-import { orderState } from "@/src/states/orderState";
-import { userState } from "@/src/states/userState";
-import { room } from "@/src/types/accommodationDetail";
-import countDays from "@/src/utils/countDays";
-import englishToKoreanFormat from "@/src/utils/englishToKoreanFormat";
-import numberFormat from "@/src/utils/numberFormat";
+import instance from "@/api/instanceApi";
+import CommonBadge from "@/components/commonBadge/CommonBadge";
+import CommonButton from "@/components/commonButton/CommonButton";
+import CommonToastLayout from "@/components/commonToast/CommonToastLayout";
+import { filterState } from "@/states/filterState";
+import { orderState } from "@/states/orderState";
+import { userState } from "@/states/userState";
+import { room } from "@/types/accommodationDetail";
+import countDays from "@/utils/countDays";
+import englishToKoreanFormat from "@/utils/englishToKoreanFormat";
+import numberFormat from "@/utils/numberFormat";
 import { format } from "date-fns";
 import _debounce from "lodash/debounce";
 import { useEffect, useState } from "react";
