+++ src/pages/basket/Basket.tsx
@@ -1,33 +1,17 @@
-import instance from "@/src/api/instanceApi";
-import CommonToastLayout from "@/src/components/commonToast/CommonToastLayout";
-import { orderState } from "@/src/states/orderState";
-import numberFormat from "@/src/utils/numberFormat";
 import { AxiosError } from "axios";
 import { useEffect, useState } from "react";
+import { CiShoppingBasket } from "react-icons/ci";
 import { useNavigate } from "react-router-dom";
 import { useSetRecoilState } from "recoil";
 import { CommonButton } from "../../components";
 import "./basket.scss";
 import SelectedAccomodation from "./selectedAccomodation/SelectedAccomodation";
-import { CiShoppingBasket } from "react-icons/ci";
-export interface RoomDescriptionType {
-  cartItemId: number;
-  checkInTime: string;
-  checkOutTime: string;
-  endDate: string;
-  headCount: number;
-  maxHeadCount: number;
-  price: number;
-  roomId: number;
-  roomName: string;
-  startDate: string;
-}
-
-export interface CartItemType {
-  accommodationName: string;
-  accommodationId: number;
-  rooms: RoomDescriptionType[];
-}
+import { CartItemType } from "@/types/basket";
+import { orderState } from "@/states/orderState";
+import CommonToastLayout from "@/components/commonToast/CommonToastLayout";
+import { calculateTotalRoomsAndPrice } from "@/utils/calculateTotalRoomsAndPrice";
+import instance from "@/api/instanceApi";
+import { flattenCartItemsData } from "@/utils/flattenCartItemsData";
 
 interface ApiResponseType {
   data: { cartItemResponseList: CartItemType[] };
@@ -36,10 +20,10 @@ interface ApiResponseType {
 
 const Basket = () => {
   const [cartItems, setCartItems] = useState<CartItemType[]>([]);
+  const navigate = useNavigate();
   const setOrderData = useSetRecoilState(orderState);
   const { showToast, ToastContainer } = CommonToastLayout();
-
-  const navigate = useNavigate();
+  const totalData = calculateTotalRoomsAndPrice(cartItems);
 
   const getCartItems = async () => {
     try {
@@ -50,6 +34,7 @@ const Basket = () => {
       console.error(axiosError);
     }
   };
+
   const deleteCartItem = async (cartItemId: number) => {
     try {
       const { data } = await instance.delete<ApiResponseType>(
@@ -80,41 +65,8 @@ const Basket = () => {
     }
   };
 
-  const calculateTotalRoomsAndPrice = (
-    data: CartItemType[]
-  ): { totalRoomsConunt: number; totalPrice: string } => {
-    let totalRoomsConunt = 0;
-    let price = 0;
-
-    data.forEach((accommodation: CartItemType) => {
-      accommodation.rooms.forEach((room: { price: number }) => {
-        totalRoomsConunt++;
-        price += room.price;
-      });
-    });
-    const totalPrice = numberFormat(price);
-    return { totalRoomsConunt, totalPrice };
-  };
-
-  const { totalRoomsConunt, totalPrice } =
-    calculateTotalRoomsAndPrice(cartItems);
-
-  const flattenData = (data: any[]) => {
-    return data.reduce((flattenedData, accommodation) => {
-      const { accommodationName, rooms } = accommodation;
-
-      return [
-        ...flattenedData,
-        ...rooms.map((room: RoomDescriptionType) => ({
-          accommodationName,
-          ...room,
-        })),
-      ];
-    }, []);
-  };
-
   const handlePaymentClick = () => {
-    const flattenedData = flattenData(cartItems);
+    const flattenedData = flattenCartItemsData(cartItems);
     setOrderData(flattenedData);
     navigate("/order?cart=true");
   };
@@ -143,9 +95,9 @@ const Basket = () => {
           />
         )}
       </div>
-      {cartItems && cartItems.length !== 0 ? (
+      {cartItems.length !== 0 ? (
         <div className="basket-container__body">
-          {cartItems?.map((item: CartItemType, index: number) => (
+          {cartItems.map((item: CartItemType, index: number) => (
             <section key={item.accommodationId}>
               <SelectedAccomodation
                 accomdationItems={item}
@@ -159,12 +111,12 @@ const Basket = () => {
           <div className="basket-container__bottom">
             <div className="total-info">
               <span className="text-subtitle5 total-info__count">
-                총 {totalRoomsConunt}건
+                총 {totalData.totalRoomsConunt}건
               </span>
               <span className="text-subtitle3">{}</span>
             </div>
             <CommonButton
-              text={`${totalPrice}원 결제하기`}
+              text={`${totalData.totalPrice}원 결제하기`}
               buttonSize="large"
               onClick={handlePaymentClick}
             />
