+++ src/pages/index.ts
@@ -1,8 +1,8 @@
-export { default as Basket } from "@/src/pages/basket/Basket";
-export { default as Home } from "@/src/pages/home/Home";
-export { default as Accommodation } from "@/src/pages/accommodation/Accommodation";
-export { default as Order } from "@/src/pages/order/Order";
-export { default as Members } from "@/src/pages/members/Members";
-export { default as OrderList } from "@/src/pages/orderList/OrderList";
-export { default as Signup } from "@/src/pages/users/signup/Signup";
-export { default as Login } from "@/src/pages/users/login/Login";
+export { default as Basket } from "@/pages/basket/Basket";
+export { default as Home } from "@/pages/home/Home";
+export { default as Accommodation } from "@/pages/accommodation/Accommodation";
+export { default as Order } from "@/pages/order/Order";
+export { default as Members } from "@/pages/members/Members";
+export { default as OrderList } from "@/pages/orderList/OrderList";
+export { default as Signup } from "@/pages/users/signup/Signup";
+export { default as Login } from "@/pages/users/login/Login";
